minor refactors and renames
