+++ tests/mock_navitia/__init__.py
@@ -31,7 +31,7 @@ import vj_6113
 import vj_6112
 import vj_6114
 import vj_96231
-import vj_80426
+import vj_840426
 import vj_R_vj1
 import vj_R_vj2
 import vj_pass_midnight
@@ -45,7 +45,7 @@ mocks = [
     vj_6113.response,
     vj_6114.response,
     vj_96231.response,
-    vj_80426.response,
+    vj_840426.response,
     vj_R_vj1.response,
     vj_R_vj2.response,
     vj_pass_midnight.response,
